refactor: isolate torrent details tabs
