+++ src/renderer/app/directives/torrent-details-info-tab/torrent-details-info-tab.directive.ts
@@ -1,0 +1,17 @@
+import { IDirective, IDirectiveFactory } from "angular";
+import { TorrentDetailsInfoTabController } from "./torrent-details-info-tab.controller";
+import html from "./torrent-details-info-tab.template.html";
+
+export class TorrentDetailsInfoTabDirective implements IDirective {
+  template = html;
+  restrict = "E";
+  scope = {
+    sections: "<",
+  };
+  controller = TorrentDetailsInfoTabController;
+  controllerAs = "ctl";
+
+  static getInstance(): IDirectiveFactory {
+    return () => new TorrentDetailsInfoTabDirective();
+  }
+}
