+++ src/renderer/app/directives/torrent-details-panel/torrent-details-panel.template.html
@@ -41,139 +41,18 @@
       </div>
     </div>
 
-    <div
-      class="torrent-details-info"
-      data-role="torrent-details-info"
+    <torrent-details-info-tab
       ng-if="!ctl.showStateMessage() && ctl.isActiveTab('info')"
-    >
-      <div class="torrent-details-section" ng-repeat="section in ctl.scope.panel.info.sections track by section.id">
-        <div class="ui tiny header">{{ section.title }}</div>
-        <table class="ui very basic compact table">
-          <tbody>
-            <tr
-              ng-repeat="field in section.fields track by field.id"
-              data-role="torrent-details-field"
-              data-field-id="{{ field.id }}"
-            >
-              <td class="label">{{ field.label }}</td>
-              <td class="value" ng-class="{ multiline: field.multiline }">{{ ctl.formatFieldValue(field) }}</td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-    </div>
+      sections="ctl.scope.panel.info.sections"
+    ></torrent-details-info-tab>
 
-    <div
-      class="torrent-details-files"
-      data-role="torrent-details-files"
+    <torrent-details-files-tab
       ng-if="!ctl.showStateMessage() && ctl.isActiveTab('files')"
-    >
-      <div class="torrent-details-files-content">
-        <div class="ui tiny negative message torrent-details-files-error" ng-if="ctl.scope.selectionError">
-          {{ ctl.scope.selectionError }}
-        </div>
-        <table
-          id="torrentDetailsFilesTable"
-          data-role="torrent-details-files-table"
-          rz-table
-          rz-profile="ctl.scope.resizeProfile"
-          rz-mode="ctl.scope.resizeMode"
-          rz-container=".torrent-details-files-content"
-          columns="ctl.scope.panel.files.columns"
-          class="ui compact single line unstackable fixed torrent resize table"
-        >
-          <thead>
-            <tr
-              sorting="ctl.changeSorting"
-              mode="ctl.scope.resizeProfile"
-              default-sort-key="name"
-              default-sort-order="false"
-              sort-key-prefix="torrent-details-files.sort-key"
-              sort-order-prefix="torrent-details-files.sort-desc"
-            >
-              <th
-                ng-repeat="column in ctl.scope.panel.files.columns track by column.id"
-                rz-col="column.id"
-                sort="column.id"
-                sort-disabled="column.sortable === false"
-                ng-class="{ 'torrent-details-file-selection-column': column.id === 'wanted' }"
-              >
-                <div ng-if="column.id === 'wanted'" class="ui checkbox torrent-details-file-checkbox">
-                  <input
-                    type="checkbox"
-                    aria-label="Select all files"
-                    data-role="torrent-details-files-select-all"
-                    ng-checked="ctl.allFilesWanted()"
-                    indeterminate-value="ctl.allFilesSelectionIndeterminate()"
-                    ng-disabled="!ctl.canSelectFiles() || ctl.scope.selectionUpdating"
-                    ng-click="ctl.toggleAllFiles($event)"
-                  />
-                  <label></label>
-                </div>
-                <span ng-if="column.id !== 'wanted'" class="torrent-details-files-header-label">{{ column.label }}</span>
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-            <tr
-              ng-repeat="row in ctl.scope.sortedFiles track by row.key"
-              ng-class="{ 'torrent-details-folder-row': row.isDirectory }"
-              data-file-index="{{ row.file.index }}"
-              data-file-path="{{ row.data.path }}"
-            >
-              <td ng-repeat="column in ctl.scope.panel.files.columns track by column.id" data-col="{{ column.id }}">
-                <div ng-if="column.id === 'wanted'" class="torrent-details-file-cell">
-                  <div class="ui checkbox torrent-details-file-checkbox">
-                    <input
-                      type="checkbox"
-                      aria-label="{{ row.isDirectory ? 'Select folder ' : 'Select file ' }}{{ row.data.name }}"
-                      data-role="torrent-details-file-checkbox"
-                      ng-checked="ctl.rowWanted(row)"
-                      indeterminate-value="ctl.rowSelectionIndeterminate(row)"
-                      ng-disabled="!ctl.canSelectFiles() || ctl.scope.selectionUpdating"
-                      ng-click="ctl.toggleRowSelection(row, $event)"
-                    />
-                    <label></label>
-                  </div>
-                </div>
-                <div ng-if="column.id !== 'wanted'" ng-switch="column.format" class="torrent-details-file-cell">
-                  <div ng-switch-when="progress" class="torrent-details-file-progress">
-                    <div class="ui tiny indicating progress">
-                      <div class="bar" ng-style="{ width: ctl.fileProgressPercent(row.data) + '%' }"></div>
-                    </div>
-                    <span>{{ ctl.fileProgressPercent(row.data) | number:1 }}%</span>
-                  </div>
-                  <div
-                    ng-switch-default
-                    ng-if="column.id === 'name'"
-                    class="torrent-details-file-name"
-                    ng-style="{ 'padding-left': (row.depth * 18) + 'px' }"
-                    title="{{ row.data.path }}"
-                    ng-click="row.isDirectory && ctl.toggleFolder(row, $event)"
-                  >
-                    <button
-                      ng-if="row.isDirectory"
-                      type="button"
-                      class="ui icon basic button torrent-details-folder-toggle"
-                      aria-label="{{ ctl.isFolderExpanded(row) ? 'Collapse' : 'Expand' }} {{ row.data.name }}"
-                      aria-expanded="{{ ctl.isFolderExpanded(row) }}"
-                      ng-click="ctl.toggleFolder(row, $event)"
-                    >
-                      <i class="icon" ng-class="ctl.isFolderExpanded(row) ? 'caret down' : 'caret right'"></i>
-                    </button>
-                    <i ng-if="!row.isDirectory" class="file outline icon torrent-details-file-icon"></i>
-                    <span>{{ row.data.name }}</span>
-                  </div>
-                  <span ng-switch-default ng-if="column.id !== 'name'">{{ ctl.formatFileCell(column, row.data) }}</span>
-                </div>
-              </td>
-            </tr>
-            <tr ng-if="!ctl.scope.sortedFiles.length">
-              <td colspan="{{ ctl.scope.panel.files.columns.length || 1 }}">No files available.</td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-    </div>
+      files="ctl.scope.panel.files"
+      resize-mode="ctl.scope.resizeMode"
+      resize-profile="ctl.scope.resizeProfile"
+      can-select-files="ctl.canSelectFiles()"
+      update-selection="ctl.updateFileSelection(files, wanted)"
+    ></torrent-details-files-tab>
   </div>
 </div>
