+++ src/renderer/styles/app/partials/layout.less
@@ -713,6 +713,13 @@ torrent-sidebar + .main-panel {
             padding: 0;
         }
 
+        torrent-details-info-tab,
+        torrent-details-files-tab {
+            display: flex;
+            flex: 1 1 auto;
+            min-height: 0;
+        }
+
         .torrent-details-state {
             display: flex;
             flex: 1 1 auto;
